+++ packages/experimental-app-router/src/index.tsx
@@ -1,1 +1,4 @@
-export { getClient } from './client.js';
+export { getClient, getAuthClient } from './client.js';
+export { faustRouteHandler } from './server/routeHandler/index.js';
+export { fetchAccessToken } from './server/auth/fetchAccessToken.js';
+export { onLogout } from './server-actions/logoutAction.js';